requirements.txt
Flask==3.1.0
llama_cpp_python==0.3.8
gunicorn==23.0.0
numpy==2.2.3
onnxruntime==1.21.0
dotenv==0.9.9
python-dotenv==1.0.1
tokenizers==0.21.1
